// backend/src/api/client/ApiTypes.ts
export type StrengthSlug =
  | 'carefulness'
  | 'compassion'
  | 'courage'
  | 'creativity'
  | 'curiosity'
  | 'enthusiasm'
  | 'fairness'
  | 'forgiveness'
  | 'gratitude'
  | 'grit'
  | 'honesty'
  | 'hope'
  | 'humour'
  | 'judgement'
  | 'kindness'
  | 'leadership'
  | 'love'
  | 'loveOfBeauty'
  | 'loveOfLearning'
  | 'modesty'
  | 'perseverance'
  | 'perspective'
  | 'selfRegulation'
  | 'socialIntelligence'
  | 'spirituality'
  | 'teamwork';
// ...
export type GroupGamePlayer = {
  id: string;
  nickname: string;
  color: string;
  avatar: string;
};
// ...
export type CreateSprintStrengthParameters = {
  id: string;
};
export type CreateSprintStrengthResponse = {
  to: string;
  strength: StrengthSlug;
};
export type CreateSprintStrengthRequest = {
  to: string;
  strength: StrengthSlug;
};
// ...
export type PatchHostSprintEvent = {
  players: GroupGamePlayer[];
  sharedStrengths: Array<{
    strength: StrengthSlug;
    from: string;
    to: string;
  }>;
  isCompleted: boolean;
  isEnded: boolean;
  updatedAt: string;
};
export type PatchPlayerSprintEvent = {
  isStarted: boolean;
  isCompleted: boolean;
  isEnded: boolean;
  players: string[];
  updatedAt: string;
};